\name{decath}
\alias{decath}
\docType{data}
\title{ data frame with decathlon results }

\description{ give results of athletes in decathlon }

\usage{data(decath)}

\format{
A data frame with 41 rows and ten columns: the ten columns correspond to the performance of the athletes for the ten events of the decathlon.
 \describe{

   \item{\code{100m}}{a numeric giving the time to run 100m}
   \item{\code{long jump}}{a numeric giving the length in long jump}
   \item{\code{shot put}}{a numeric giving the length in shot put}
   \item{\code{high jump}}{a numeric giving the high in high jump}
   \item{\code{400m}}{a numeric giving the time to run 400m}
   \item{\code{110m hurdles}}{a numeric giving the time to run 110m hurdles}
   \item{\code{discus}}{a numeric giving the time to run 100m}
   \item{\code{pole vault}}{a numeric giving the high in pole vault}
   \item{\code{javeline}}{a numeric giving the length in javeline}
   \item{\code{1500m}}{a numeric giving the time to run 1500m}

          }
        }

\keyword{datasets}
