\name{GCCA}
\alias{GCCA}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
function performing a Generalized Canonical Correlation Analysis on a given data table or on (simulated) data streams.
}
\description{
This function permits user to perform fast Generalized Canonical Correlation Analysis on (high dimensional on-line) data.  

Using stochastic processes, several estimation may be calculated depending on the user goals:
principal canonical factors, principal canonical axes, eigenvalues and correlation coefficients between the original
variables and the principal canonical factors can be computed and updated recursively.
 
Moreover, graphics may be printed in order to give a better visualization and meaning/interpretation of
the data. It's possible to get meanningful 2D visualizations: the observations are projected on the plans generated
 by direction vectors of the principal axes.  Correlation circles that show a
projection of the initial variables in the factors space are also available.

Reminding that GCCA is an analysis achieved on q groups of variables (q > or = 2), we precise that the \code{GCCA} function computes the general elements
(principal factors, principal axes, ...) . If the user needs the canonical elements associated with the $k^{th}$ group (k=1,\dots,q), it's possible to deduce them from the
 general elements: the principal canonical factors associated with the $k^{th}$ group are 
 colinears to the vectors composed of the relevant components of the principal general factors.
}
\usage{
GCCA(data,groups,stream=TRUE,nb_fact,principal_factors=FALSE,principal_axes=TRUE,
eigenvalues=FALSE,corr=FALSE,graphics=FALSE,data_init,exec_time,print_step)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{data}{
matrix : data to be analysed. The columns will represent the variables and 
the rows will represent the observations (sometimes called records, subjects or cases).
If you consider on-line data (\code{stream=TRUE}), of course all the observations are not available at the
beginning of the analysis. Thus, each observation (line of the matrix) will be treated 
individually (and then forgotten) to simulate the data stream.     
}
  \item{groups}{vector : each component of \code{groups} gives the length of the associated group of variables.
}
  \item{stream}{
boolean : stream=TRUE if you consider data streams; stream=FALSE otherwise.
}
  \item{nb_fact}{
integer : number of elements to be calculated. For instance, if you need
 the three first principal general factors, nb_fact=3
}
  \item{principal_factors}{ boolean : principal_factors=TRUE if you need the value of some first
principal general factors; principal_factors=FALSE otherwise.
}
  \item{principal_axes}{ boolean : principal_axes=TRUE if you need the value of some first
principal general axes; principal_axes=FALSE otherwise.
}
  \item{eigenvalues}{boolean : eigenvalues=TRUE if you need the value of some first
eigenvalues; eigenvalues=FALSE otherwise.
}
  \item{corr}{boolean : corr=TRUE if you need the value of correlation coefficient between the
original variables and some first principal canonical factors; corr=FALSE otherwise.
}
  \item{graphics}{
boolean : if graphics=TRUE, 2D visualizations will be plotted, representing
projections of the observations on all the plans generated by the \code{nb_fact} direction vectors
of the principal axes. If corr=TRUE, correlation circles that show a projection of the initial variables in the factors
spaces will also be plotted; if graphics=FALSE, no graphics will be plotted.


N.B. : For the fist kind of graphics, a gradation of grey was chosen to represent the "age" of the observations. The darker a point is, the newer the observation is. This coding
is meaningful for a data stream (stream=TRUE). if stream=FALSE, the "age" concept is less relevant: in this configuration, the darker a point is, the lower the observation is in
the matrix (database).  
}
  \item{data_init}{
integer : this argument is only relevant when stream=TRUE. It represents the number
of data used to intialize the analysis. It must be at least equal to the  number of columns of the dataset.
}
  \item{exec_time}{
real : execution time. If stream=TRUE, the analysis will stop when one of
these two events happen: all the data have been taken into account or execution time achieved. Otherwise, it's the execution time of \code{\link{GCCA_iter}}.
}
  \item{print_step}{
integer : this argument is only relevant when stream=TRUE. Each time that a new observation
is taken into account, a new estimation of the principal general factors (or principal canonical axes, ...) is computed but not
necessarily printed: it will be printed only every \code{print_step} observations.  
}
}
\details{
GCCA is a generalization of CCA in the sense that it performs an analysis on more than 2 groups of quantitative variables.
}
\value{
results that user needs (principal factors, principal axes, eigenvalues and correlation coefficients)
}
\references{
Stochastic approximation of the factors of a generalized canonical correlation analysis, Monnez J-M., Statistics and Probability Letters 78, 14 (2008) 2210-2216
}
\author{
BAR Romain, Université de Lorraine, IECL, INRIA : BIGS group.
}
\seealso{ 
\code{\link{GCCA_iter}},
\code{\link{CCA}},
\code{\link{CCA_iter}}
}
\examples{

data(vins_de_loire)
GCCA(rbind(vins_de_loire,vins_de_loire,vins_de_loire),c(5,3,10,9,2),stream=TRUE,3,
principal_factors=TRUE,principal_axes=TRUE,eigenvalues=TRUE,corr=TRUE,30,graphics=TRUE,
1,print_step=10)

}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{generalized canonical correlation analysis}
