\name{norm}
\alias{norm}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
computes the norm of a matrix in the Hilbert-Schmidt sense
}
\description{
computes the norm of a matrix in the Hilbert-Schmidt sense
}
\usage{
norm(A)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{A}{matrix
}
}

\value{
norm of A
}
\references{
http://mathworld.wolfram.com/Hilbert-SchmidtNorm.html
}
\author{
Romain Bar, Université de Lorraine, IECL, INRIA: BIGS group
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ Hilbert-Schmidt norm }
