\name{preprocess_CA}
\alias{preprocess_CA}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
preprocessing for Correpondence Analysis
}
\description{
performs an encoding of the original data set before launching CA
}
\usage{
preprocess_CA(df)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{df}{matrix : data set before encoding
}
}

\value{
matrix : data set after encoding
}

\author{
Romain Bar, Université de Lorraine, IECL, INRIA : BIGS group
}
\seealso{ 
\code{\link{preprocess_CDA}},
\code{\link{preprocess_MCA}}
}



% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{correpondence analysis }
\keyword{ encoding }
