\name{psy}
\alias{psy}
\docType{data}
\title{ Psychological data }

\description{ Investigation about the associations between psychological measures and academic achievement measures. }

\usage{data(psy)}

\format{
A data frame with 600 observations (rows) on the following 7 quantitative variables (columns) divided in 2 groups: 3 psychological variables and 4 academic variables.
 \describe{
   \item{\code{locus_of_control}}{a numeric giving the "measurement" of locus of control}
   \item{\code{self_concept}}{a numeric giving the "measurement" of self concept}
   \item{\code{motivation}}{a numeric giving the "measurement" of motivation}
   \item{\code{read}}{a numeric giving the result on a standardized test in reading}
   \item{\code{write}}{a numeric giving the result on a standardized test in writing}
   \item{\code{math}}{a numeric giving the result on a standardized test in mathematics}
   \item{\code{science}}{a numeric giving the result on a standardized test in sciences}
          }
        }

\keyword{datasets}
