\name{vins_de_loire}
\alias{vins_de_loire}
\docType{data}
\title{ Wine from Loire }

\description{ give sensory descriptions of 21 wines from Loire }

\usage{data(vins_de_loire)}

\format{
A data frame with 21 rows (the number of wines) and 29 columns: quantitative variables corresponding to sensory descriptors.
 \describe{
   \item{\code{Odor.Intensity.before.shaking}}{a numeric (note between 0 and 5) giving the odor intensity of the wine before shaking}
   \item{\code{Aroma.quality.before.shaking}}{a numeric (note between 0 and 5) giving the aroma quality of the wine before shaking}  
   \item{\code{Fruity.before.shaking}}{a numeric (note between 0 and 5) giving the fruity odor intensity of the wine before shaking}
   \item{\code{Flower.before.shaking}}{a numeric (note between 0 and 5) giving the flower odor intensity of the wine before shaking}
   \item{\code{Spice.before.shaking}}{a numeric (note between 0 and 5) giving the spicy odor intensity of the wine before shaking}
   \item{\code{Visual.intensity}}{a numeric (note between 0 and 5) giving the visual intensity of the wine}
   \item{\code{Nuance}}{a numeric (note between 0 and 5) giving the nuance quality of the wine}
   \item{\code{Surface.feeling}}{a numeric (note between 0 and 5) giving the surface feeling of the wine}
   \item{\code{Odor.Intensity}}{a numeric (note between 0 and 5) giving the odor intensity of the wine} 
   \item{\code{Quality.of.odour}}{a numeric (note between 0 and 5) giving the quality of odour of the wine}
   \item{\code{Fruity}}{a numeric (note between 0 and 5) giving the fruity odor intensity of the wine}
   \item{\code{Flower}}{a numeric (note between 0 and 5) giving the flower odor intensity of the wine}
   \item{\code{Spice}}{a numeric (note between 0 and 5) giving the spicy odor intensity of the wine}
   \item{\code{Plante}}{a numeric (note between 0 and 5) giving the plante odor intensity of the wine}
   \item{\code{Phenolic}}{a numeric (note between 0 and 5) giving the phenolic odor intensity of the wine}   
   \item{\code{Aroma.intensity}}{a numeric (note between 0 and 5) giving the aroma intensity of the wine}
   \item{\code{Aroma.persistency}}{a numeric (note between 0 and 5) giving the aroma persistency of the wine}
   \item{\code{Aroma.quality}}{a numeric (note between 0 and 5) giving the aroma quality of the wine}
   \item{\code{Attack.intensity}}{a numeric (note between 0 and 5) giving the attack intensity of the wine}
   \item{\code{Acidity}}{a numeric (note between 0 and 5) giving the acidity intensity of the wine}
   \item{\code{Astringency}}{a numeric (note between 0 and 5) giving the astringency intensity of the wine}
   \item{\code{Alcohol}}{a numeric (note between 0 and 5) giving the alcohol intensity of the wine} 
   \item{\code{Balance}}{a numeric (note between 0 and 5) giving the balance of the wine}
   \item{\code{Smooth}}{a numeric (note between 0 and 5) giving the smoothness of the wine}
   \item{\code{Bitterness}}{a numeric (note between 0 and 5) giving the bitterness of the wine}
   \item{\code{Intensity}}{a numeric (note between 0 and 5) giving the intensity of the wine}
   \item{\code{Harmony}}{a numeric (note between 0 and 5) giving the harmony of the wine}
   \item{\code{Overall.quality}}{a numeric (note between 0 and 5) giving the overall feeling about the wine}
   \item{\code{Typical}}{a numeric (note between 0 and 5) giving the typicity of the wine} 
       }
        }

\keyword{datasets}

