\name{wolves}
\alias{wolves}
\docType{data}
\title{ Study on skull wolves }

\description{ Skull morphometric data on Rocky Mountain and Arctic wolves (Canis Lupus L.) taken from Morrison
(1990), originally from Jolicoeur (1959).}

\usage{data(wolves)}

\format{
A data frame with 25 observations on the following 10 variables.

\describe{

\item{\code{V1}}{ palatal length, a numeric vector}
\item{\code{V2}}{ postpalatal length, a numeric vector}
\item{\code{V3}}{ zygomatic width, a numeric vector}
\item{\code{V4}}{ palatal width outside first upper molars, a numeric vector}
\item{\code{V5}}{ palatal width inside second upper molars, a numeric vector}
\item{\code{V6}}{ postglenoid foramina width, a numeric vector}
\item{\code{V7}}{ interorbital width, a numeric vector}
\item{\code{V8}}{ braincase width, a numeric vector}
\item{\code{V9}}{ crown length, a numeric vector}
\item{\code{V10}}{ a factor with levels ar:f ar:m rm:f rm:m, comprising the combinations of location and
sex}
 
        }
        }
\keyword{datasets}
